Provide a method to build both static and dynamic libraries
In the past both wcecompat.lib and wcecompatex.lib have been created,
both of which are identical static libraries.

With this change the wcecompat.lib corresponds with wcecompat.dll,
a dynamic library; wcecompatex.lib is still a static library;
and winmain.lib provides winmain.obj.

Note that winmain.obj is not part of the dynamic library.

Other changes include:
- remove defintions for _wceread and _wcewrite
- create wcecompatex.pdb
- improve CFLAGS for the variety of WinCE targets supported

Issue:IA-151
Change-Id: I8ff1bef6364bf5b1b1906aa335551f4d9f7ada77